requirements.txt
mujoco==3.2.2
trimesh==4.4.1
hydra-core==1.3.2
rtree==1.2.0
plotly==5.22.0
tqdm==4.66.4
scipy==1.14.0
pandas==2.2.2
opencv-python==4.10.0.84
